Download both MODIS files for VDM

+++ vdi.js
@@ -3,25 +3,26 @@ const cheerio = require("cheerio");
 const fs = require("fs");
 const { exec } = require("child_process");
 
+const token =
+  "eyJ0eXAiOiJKV1QiLCJvcmlnaW4iOiJFYXJ0aGRhdGEgTG9naW4iLCJzaWciOiJlZGxqd3RwdWJrZXlfb3BzIiwiYWxnIjoiUlMyNTYifQ.eyJ0eXBlIjoiVXNlciIsInVpZCI6Iml2by5taW5pYyIsImV4cCI6MTY5MjM0ODQxOSwiaWF0IjoxNjg3MTY0NDE5LCJpc3MiOiJFYXJ0aGRhdGEgTG9naW4ifQ.1OT1RZv5jAFEi0NfIAAzXg-5DN_qZaG1OC2HOHynL-0Qnb4Xlx4AxZiJb3pn4JUJRH6VDWnY6P4cf4DumWdacS7sxk5GIRBsFDYV_uYHTHIEFFhijgVEydyq6J5WRr55nG9blVv8n9bnuM3RBoz4C1XmGAxsL6nAgzwWg50nItc0MtIfN3jCurf96lR8BqWxPewn9_ZUAdNE7hQYYBDYs9hCwc5UfWVJgznYHsHRoQpUzklB2EzqMG0ajEdtTE-mJi07BIbubtxGIXngDWHjlkP3tKYPLSeS66rewfP8XfNi0-NiicHiSIBtZ_icbkbrxDz-W5zNI9rJSXIEeiRPGg";
+let url = "https://e4ftl01.cr.usgs.gov/MOLT/MOD13Q1.061/";
+let fileUrl = "";
+let originalFilename = "";
+let blnExistFile = false; //Need to stop iterating when this value is true
+let blnFound = false; //Same as above, but for MOD09GA.061
+
 const fileIdentifier = ".h19v04.";
 const fileExtension = ".hdf";
 const downloadFolder = "C:/Users/PC/Desktop/modis/vdi/hdf/";
-let downloadedFileName = "";
+
+let urlModis = "https://e4ftl01.cr.usgs.gov/MOLT/MOD09GA.061/";
+let originalModisFilename = "";
+const downloadModisFolder = "C:/Users/PC/Desktop/modis/ndii/";
 
 //TODO: Check if file is already downloaded.
 //TODO: Create path regarding selected date.
 //TODO: Rename file by format that JICA sent.
 //TODO: Log result of the action.
-
-//let url = "https://e4ftl01.cr.usgs.gov/MOLT/MOD13Q1.061/2023.01.17/";
-let url = "https://e4ftl01.cr.usgs.gov/MOLT/MOD13Q1.061/";
-
-const token =
-  "eyJ0eXAiOiJKV1QiLCJvcmlnaW4iOiJFYXJ0aGRhdGEgTG9naW4iLCJzaWciOiJlZGxqd3RwdWJrZXlfb3BzIiwiYWxnIjoiUlMyNTYifQ.eyJ0eXBlIjoiVXNlciIsInVpZCI6Iml2by5taW5pYyIsImV4cCI6MTY2OTQ0ODkyOCwiaWF0IjoxNjY0MjY0OTI4LCJpc3MiOiJFYXJ0aGRhdGEgTG9naW4ifQ.phOmuKNV1-VypNsLKb-odPiWYC2VUPuQNLYnLnSuwjyzI8-Cs8KuYbEwPVu7TVsiGWdahaayPOnRuewv0rlJw0WQ7v6Pq5C_dUZOCZS2GEjwCiwk4dXTTjt9SmGIeTVLsvhVtWQHHGI-q4Yk7gW0EHJ0vdP5IeViJvg2eWd5Hqi1qRG4sic9KNwM8ZWLYiND4gBQ5ZfSIxbKwAYAK-Z07rWkzPgP4fNCEe43K-tksEc5q59zOzP7dG0IlWA1iKHc8DlJR2QrsgvFYQUBF9NQ4Q_POpaRkzILmqmwRGVqpuKp86wQ1bjHRm9C53D8vtqisgmBrpJjul72aXha0M6smg";
-let fileUrl = "";
-
-let originalFilename = "";
-let blnExistFile = false; //Need to stop iterating when this value is true
 
 async function scrapeData(url) {
   try {
@@ -34,12 +35,12 @@ async function scrapeData(url) {
         console.log("FileName", originalFilename);
         fileUrl = url + originalFilename;
         blnExistFile = true;
-        downloadedFileName = originalFilename;
       }
     });
     if (blnExistFile) {
       console.log("URL: ", fileUrl);
-      //await downloadFile(fileUrl, token, downloadFolder + originalFilename);
+      await downloadFile(fileUrl, token, downloadFolder + originalFilename);
+      await getLatestModis09A1();
     }
   } catch (err) {
     //console.error("Error", err);
@@ -47,15 +48,17 @@ async function scrapeData(url) {
 }
 
 async function downloadFile(url, token, filePath) {
-  let command = `wget -e robots=off -m -np -R .html,.tmp -nH --cut-dirs=4 "${url}" --header "Authorization: Bearer ${token}" -O ${filePath}`;
-  exec(command, (err, stdout, stderr) => {
-    if (err) {
-      console.log("ERROR", err);
-      return;
-    }
-    console.log(`stdout: ${stdout}`);
-    console.log(`stderr: ${stderr}`);
-  });
+  if (!fs.existsSync(filePath)) {
+    let command = `wget -e robots=off -m -np -R .html,.tmp -nH --cut-dirs=4 "${url}" --header "Authorization: Bearer ${token}" -O ${filePath}`;
+    exec(command, (err, stdout, stderr) => {
+      if (err) {
+        console.log("ERROR", err);
+        return;
+      }
+      console.log(`stdout: ${stdout}`);
+      console.log(`stderr: ${stderr}`);
+    });
+  }
 }
 
 //scrapeData(url);
@@ -78,4 +81,39 @@ function padTo2Digits(num) {
 
 function formatDate(date) {
   return [date.getFullYear(), padTo2Digits(date.getMonth() + 1), padTo2Digits(date.getDate())].join(".");
+}
+
+/** Methods for MODIS09A1 - download newest file and store filename value */
+async function getLatestModis09A1() {
+  let loop = new Date();
+  while (!blnFound) {
+    console.log("Date", formatDate(loop));
+    await scrapeDataModis09A1(urlModis + formatDate(loop) + "/");
+
+    let newDate = loop.setDate(loop.getDate() - 1);
+    loop = new Date(newDate);
+  }
+}
+
+async function scrapeDataModis09A1(tempUrl) {
+  try {
+    const { data } = await axios.get(tempUrl);
+    const $ = cheerio.load(data);
+    const listItems = $("a");
+    listItems.each((idx, el) => {
+      if ($(el).text().includes(fileIdentifier) && $(el).text().endsWith(fileExtension)) {
+        originalModisFilename = $(el).text();
+        console.log("FileName09A", originalModisFilename);
+        blnFound = true;
+        fileUrl = tempUrl + originalModisFilename;
+        originalModisFilename = originalModisFilename;
+      }
+    });
+    if (blnFound) {
+      console.log("URL09A: ", tempUrl);
+      await downloadFile(fileUrl, token, downloadModisFolder + originalModisFilename);
+    }
+  } catch (err) {
+    //console.error("Error", err);
+  }
 }
